+++ test/type_generator-test.ts
@@ -185,6 +185,7 @@ export interface FitOptions {
     tank_updates?: ParsedTankUpdate[];
     tank_summaries?: ParsedTankSummary[];
     jumps?: ParsedJump[];
+    sets?: ParsedSet[];
     splits?: ParsedSplit[];
     split_summaries?: ParsedSplitSummary[];
     time_in_zone?: ParsedTimeInZone[];